basic camera generation script (x,y,z,theta,phi)
Coordinates and the camera's intrinsic and extrinsic properties are specified in the output file (by default cameras/cameras.txt). On closer inspection, these properties are incorrect and need to be fixed on an updated version. The current process is as follows: calculate the scale of the coordinates x and z (25&26), calculate phi in terms of position facing in the circle, and save the coordinates (y and theta are constant). In actuality, the "extrinsic" and "intrinsic" properties should be used in the calculation to recover the coordinates more precisely and also recover the focal length.

All of these changes will be made in the next version.